final version v1

+++ db/schema.sql
@@ -4,5 +4,5 @@ CREATE TABLE employees (
   id SERIAL PRIMARY KEY,
   name TEXT NOT NULL,
   birthday DATE NOT NULL,
-  salary NUMERIC(10, 2) NOT NULL CHECK (salary >= 0)
+  salary INTEGER NOT NULL CHECK (salary >= 0)
 );
